Added PathEntity and Updated MarkerLineEntity
