+++ packages/mapray/src/index.js
@@ -18,6 +18,7 @@ import StandardImageProvider from "./StandardImageProvider";
 import Scene from "./Scene";
 import Entity from "./Entity";
 import MarkerLineEntity from "./MarkerLineEntity";
+import PathEntity from "./PathEntity";
 import TextEntity from "./TextEntity";
 import ModelEntity from "./ModelEntity";
 import PolygonEntity from "./PolygonEntity";
@@ -58,6 +59,7 @@ var mapray = {
     Scene,
     Entity,
     MarkerLineEntity,
+    PathEntity,
     TextEntity,
     ModelEntity,
     PolygonEntity,
